Add FileBackedCASDatabase
This adds an implementation for file-backed CAS database that isn't
particularly efficient but useful for performing tests. This also adds
support for writing and reading to file-backed CAS using `retool cas`
subtool.

The cas subtool only operates on the file-backed CAS right now but that
can be evolved over time as we implement more CAS databases.