Refactor ProviderMap management from the client developers.
This introduces a new API for clients of LLBBuildSystem to declare
dependencies for LLBConfiguredTarget implementations that allows for a
cleaner management by the build system.

It also introduces new APIs into the ruleContext to retrieve those
providers with typed methods. This effectively hides the fact that
LLBProviderMap even exists, which is better encapsulation.