Update package dependencies to build with Swift 5.7
The previous minimum version for swift-tools-support-async was really
ancient and would not build. Update it to a recent version (0.8.3).

swift-tools-support-core PR #345 is needed to build with Xcode 14.
The 0.2.7 tag includes that fix, so require that version. (Note that
the other tags from around the same point, e.g., 0.2.6 and 0.3.0,
have an additional requirement for macOS 10.13, which causes trouble
unless all transitive dependencies raise their minimum macOS version.)