Record full keys for llbuild2fx InternalValues
It has been very difficult or even impossible to determine why a requested
key was not found in the cache, e.g., when re-running a build with a few
changes. Only the key information reflected in the InternalValue.cachePath
was recorded. This change adds a reference in every InternalValue to the
full JSON-encoded key for that value. With this, you can do things like
comparing two builds to see exactly what differs in their requested keys.

This is implemented by encoding each key as JSON in the CAS and then adding
a CAS reference as the first element in the refs for an InternalValue. Note
that this is an incompatible change in the encoding of InternalValue that
needs to be versioned appropriately.