Make the action IDs more resilient to potential collisions
If the mnemonic and arguments happen to be the same for two
`LLBBuildEventActionDescription` instances, but the rest of the
action description is different, we would previously produce the
same identifier, which was used in log file names and caused issues
where users would get their files overwritten etc. This fixes that
by adding action owner, label and description into the hash.

Signed-off-by: Silas Snider <swsnider@apple.com>